\name{capitalize_first_letter}
\alias{capitalize_first_letter}
\title{Capitalization of First Letter}

\description{
Capitalizes the first letter of the word.
}

\usage{
capitalize_first_letter(x)
}

\arguments{
  \item{x}{A character vector.} 
}

\value{It ouputs a character vector of the same length.}

\examples{
capitalize_first_letter(word)
}